Create AutoTexter
This is a fully working Google script that is already in use and tested. It needs to be cleaned up a bit and optimized but works as is.

This is meant to be used with a Gsheet in Google drive as it is a script that runs in a Google spread sheet. There are some features I would like to add down the road so that the end user can edit various parameters without editing the code directly.